% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(dataset_name, path = "tmp", delete = TRUE)
}
\arguments{
\item{dataset_name}{Name of dataset. Currently accepts "human_DLPFC",
"mouse_coronal", or "seqFISH_mouseEmbryo".}

\item{path}{Directory to save temporary files. Default = "tmp".}

\item{delete}{Whether to delete temporary files. Default = TRUE.}
}
\value{
SpatialExperiment object
}
\description{
Function to load data objects from STdata package
}
\details{
Function to load data objects from Dropbox links (until package is moved to
Bioconductor's ExperimentHub).
}
\examples{
#spe <- load_data("human_DLPFC")
#spe <- load_data("mouse_coronal")
#spe <- load_data("seqFISH_mouseEmbryo")

}
