% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\alias{Visium_humanDLPFC}
\alias{Visium_mouseCoronal}
\alias{seqFISH_mouseEmbryo}
\alias{ST_mouseOB}
\alias{SlideSeqV2_mouseHPC}
\title{Functions to load data}
\usage{
Visium_humanDLPFC()

Visium_mouseCoronal()

seqFISH_mouseEmbryo()

ST_mouseOB()

SlideSeqV2_mouseHPC()
}
\value{
SpatialExperiment object
}
\description{
Named functions to load objects from URLs
}
\details{
Named functions to load objects from URLs in development version of
STexampleData package
}
\examples{
# spe <- Visium_humanDLPFC()
# spe <- Visium_mouseCoronal()
# spe <- seqFISH_mouseEmbryo()
# spe <- ST_mouseOB()
# spe <- SlideSeqV2_mouseHPC()

}
