% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(dataset_name, path = "tmp", delete = TRUE)
}
\arguments{
\item{dataset_name}{Name of dataset. Options are "Visium_humanDLPFC", 
"Visium_mouseCoronal", or "seqFISH_mouseEmbryo".}

\item{path}{Directory to save temporary files. Default = "tmp".}

\item{delete}{Whether to delete temporary files. Default = TRUE.}
}
\value{
SpatialExperiment object
}
\description{
Temporary function to load objects in devel version of STexampleData
}
\details{
Temporary function to load objects in devel version of STexampleData package
}
\examples{
#spe <- load_data("Visium_humanDLPFC")

}
