% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(data_name, path = ".", delete = TRUE)
}
\arguments{
\item{data_name}{Name of dataset. Currently accepts "human_DLPFC".}

\item{path}{Directory to save temporary file. Default = current working
directory.}

\item{delete}{Whether to delete temporary file. Default = TRUE.}
}
\value{
SingleCellExperiment or SpatialExperiment object
}
\description{
Temporary function to load data objects from Dropbox
}
\details{
Temporary function to load data objects from Dropbox before ExperimentHub
package is set up
}
\examples{
spe <- load_data("human_DLPFC")

}
