\name{STexampleData}
\docType{data}
\alias{STexampleData}
\alias{STexampleData-package}
\alias{Visium_humanDLPFC}
\alias{Visium_mouseCoronal}
\alias{seqFISH_mouseEmbryo}
\alias{ST_mouseOB}
\alias{SlideSeqV2_mouseHPC}
\title{
Collection of spatially-resolved transcriptomics datasets in SpatialExperiment 
Bioconductor format
}
\description{
Collection of spatially-resolved transcriptomics (SRT) datasets in 
SpatialExperiment Bioconductor format, for use in examples, demonstrations, 
and tutorials. The datasets are from several different SRT platforms and 
have been sourced from various publicly available sources. Several datasets 
include images and/or ground truth annotation labels.
}
\details{
The \code{STexampleData} package contains a collection of spatially-resolved 
transcriptomics (SRT) datasets, which have been formatted into the 
\code{SpatialExperiment} Bioconductor class, for use in examples, 
demonstrations, and tutorials.

The datasets are from several different SRT platforms and have been sourced 
from various publicly available sources. Several datasets include images and/or 
ground truth annotation labels.

Additional examples and documentation are provided in the package vignette.

\emph{Datasets}

The \code{STexampleData} package contains the following datasets:

\itemize{
\item Visium_humanDLPFC (10x Genomics Visium): A single sample (sample 151673) 
of human brain dorsolateral prefrontal cortex (DLPFC) in the human brain, 
measured using the 10x Genomics Visium platform. This is a subset of the full 
dataset containing 12 samples from 3 neurotypical donors, published by 
Maynard and Collado-Torres et al. (2021). The full dataset is available from 
the \code{spatialLIBD} Bioconductor package.

\item Visium_mouseCoronal (10x Genomics Visium): A single coronal section from 
the mouse brain, spanning one hemisphere. This dataset was previously released 
by 10x Genomics on their website.

\item seqFISH_mouseEmbryo (seqFISH): A subset of cells (embryo 1, z-slice 2) 
from a previously published dataset investigating mouse embryogenesis by 
Lohoff and Ghazanfar et al. (2020), generated using the seqFISH platform. The 
full dataset is available from the original publication.

\item ST_mouseOB (Spatial Transcriptomics): A single sample from the mouse 
brain olfactory bulb (OB), measured with the Spatial Transcriptomics platform 
(Stahl et al. 2016). This dataset contains annotations for five cell layers 
from the original authors.

\item SlideSeqV2_mouseHPC (Slide-seqV2): A single sample of mouse brain from 
the hippocampus (HPC) and surrounding regions, measured with the Slide-seqV2 
platform (Stickels et al. 2020). This dataset contains cell type annotations 
generated by Cable et al. (2021).
}
}
\examples{
# load using dataset name
spe <- Visium_humanDLPFC()
spe
}
